% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MWE.R
\name{find_erp_curves}
\alias{find_erp_curves}
\title{Apply findcurves() to a dataframe of ERPs and return MWEs}
\usage{
find_erp_curves(df, timep, grp, erp, alpha = 0.05)
}
\arguments{
\item{df}{dataframe, input data}

\item{timep}{numeric, timepoints to calculate for}

\item{grp}{string, participant groups in the dataframe}

\item{erp}{string, ERP ids}

\item{alpha}{numeric, value of alpha}
}
\value{
dataframe grouping findcurves dataframes by group and ERPs::
  colMeans plus upper and lower MWE confidence bands, plus 1-alpha confidence intervals
}
\description{
Wrangle dataframe to matrix and apply findcurves(), to return dataframe of MWEs by group and ERP
}
\examples{
findcurves(curvematrix, alpha = a)

}
