% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MWE.R
\name{find_mwe}
\alias{find_mwe}
\title{Finds MWE efficiently using separate validation data as documented in Korpela et al.
(2014). The algorithm has time complexity of O(m*n*log(n)), where m is the length of 
time series and n is the rows in the training data or validation data, whichever is
larger.}
\usage{
find_mwe(max_k, max_alpha, data_tr, data_va)
}
\arguments{
\item{max_k}{integer, maximum value of k searched}

\item{max_alpha}{real number, fraction of validation data samples out of MWE}

\item{data_tr}{nXm matrix, training data}

\item{data_va}{n'Xm matrix, validation data}
}
\value{
Returns a list that contains number of time series removed k, final alpha 
 (should be the largest alpha which is at most max_alpha), and lower and upper 
 bounds.
}
\description{
Finds MWE efficiently using separate validation data as documented in Korpela et al.
(2014). The algorithm has time complexity of O(m*n*log(n)), where m is the length of 
time series and n is the rows in the training data or validation data, whichever is
larger.
}
