% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MWE.R
\name{findcurves}
\alias{findcurves}
\title{Apply the find_mwe() and return dataframe of CBs and CIs}
\usage{
findcurves(data, mtr = 5000, mva = 5000, alpha = 0.05, signflip = FALSE)
}
\arguments{
\item{data}{[n,m] numeric, input data}

\item{mtr}{numeric, number of training resamples}

\item{mva}{numeric, number of validation resamples}

\item{alpha}{numeric, value of alpha}

\item{signflip}{boolean, ...}
}
\value{
dataframe of colMeans plus upper and lower MWE confidence bands, plus 1-alpha confidence intervals
}
\description{
Apply find_mwe() to a data matrix, and return dataframe of means, MWE confidence bands, and 1-alpha confidence intervals
}
\examples{
findcurves(curvematrix, alpha = a)

}
