% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5.tools.R
\name{h5.get.dataset.attr}
\alias{h5.get.dataset.attr}
\title{Fetch some attribute from a HDF5 file}
\usage{
h5.get.dataset.attr(hdf5file, dgroup, dataset, attrib)
}
\arguments{
\item{hdf5file}{string, Full filename of the HDF5 file, e.g.}

\item{dgroup}{string, Data set group name, e.g.}

\item{dataset}{string, Data set name, e.g.}

\item{attrib}{string, Data attribute name}
}
\value{
Attribute data
}
\description{
Fetch some attribute from a HDF5 file
}
\examples{
\dontrun{
hdf5file = '/home/jkor/tmp/test.h5'
dgroup <- '/erp'
dataset <- 'SEAMPILOT301'
attrib <- 'maxdim'
}

}
