% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5.tools.R
\name{h5attr2dimname}
\alias{h5attr2dimname}
\title{Convert HDF5 attribute to dimnames() compatible format}
\usage{
h5attr2dimname(attrdata)
}
\arguments{
\item{attrdata}{HDF5 attribute data, Output of h5::h5attr(object, attributename)}
}
\value{
Attribute data in dimnames() compatible format
}
\description{
Convert HDF5 attribute to dimnames() combatible format
Note: HDF5 representation of char arrays / cellstr is user defined.
Using the convention of collapsing into single string with ";"
separators.
}
