% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_utils.R
\name{load_ctap_mc_measurement}
\alias{load_ctap_mc_measurement}
\title{Load CTAP MC measurement sheet/table}
\usage{
load_ctap_mc_measurement(mc_file, measurement_sheet,
  dateformat = "\%d.\%m.\%Y")
}
\arguments{
\item{mc_file}{character, Full path to the MC file}

\item{measurement_sheet}{character, Name of sheet that lists measurements}

\item{dateformat}{character, Format for the date column}
}
\value{
A tibble with the contents of the measurement_sheet sheet/table
}
\description{
Load CTAP MC measurement sheet/table. File can be xlsx or sqlite.
}
