% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5.tools.R
\name{loaddir.h5.erp}
\alias{loaddir.h5.erp}
\title{Load ERP data from all hdf5 files in a directory, output a list}
\usage{
loaddir.h5.erp(indir, element, pattern = "*_ERPdata.h5")
}
\arguments{
\item{indir}{character, Directory name}

\item{element}{character, Data element to load}

\item{pattern}{character, File name pattern to use}
}
\value{
A list of data matrices
}
\description{
Load ERP data from all hdf5 files in a directory, output a list.
Uses loaddir.h5() and adds some metadata on top.
}
\examples{
\dontrun{
indir <- '/path/to/h5/'
pattern <- '.*NB01.*_ERPdata.h5'
element <- '/erp'
}

}
