% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtools.R
\name{mystr_detect_replace}
\alias{mystr_detect_replace}
\title{Detect substring and replace if there is a match}
\usage{
mystr_detect_replace(string, pattern, replacement = pattern)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something coercible to one.}

\item{pattern}{Vector of patterns to look for. See stringr::str_which() for details.}

\item{replacement}{Vector of replacement strings. Has to be of the same length as pattern.}
}
\value{
A character vector.
}
\description{
Replaces elements in character vector 'string' with elements found in
'replacement', if the element matches one of the patterns in
'pattern'. Matches of pattern[i] are replaced with replacement[i].

Intended usage is to convert a vector of path strings into a vector
of class labels.
}
