% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MWE.R
\name{plot_ci_one}
\alias{plot_ci_one}
\title{ggplot2 line plotting for ERP MWEs}
\usage{
plot_ci_one(
  curve_list,
  curve_colors = c("#0072B2", "#D55E00"),
  grp = NA,
  erp = NA,
  ylims = NA,
  ymirror = FALSE,
  ybreakstep = 5
)
}
\arguments{
\item{curve_list}{list, input data}

\item{curve_colors}{vector, hexadecimal strings for two curve colors for MWE and CI respectively}

\item{grp}{string, participant groups in the dataframe}

\item{erp}{string, ERP ids}

\item{ylims}{vector, set manual y limits of the plots}

\item{ymirror}{boolean, if finding ylims from data, set TRUE to have equal min/max, or FALSE to use data's range}

\item{ybreakstep}{numeric, size of y axis breaks for ticks, centred on zero}
}
\description{
ggplot2 line plotting for lsit of ERP MWE dataframes with mean, MWE confidence bands, and 1-alpha confidence intervals
}
