% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MWE.R
\name{plotlines}
\alias{plotlines}
\title{Base R line plotting useful for MWEs}
\usage{
plotlines(x, tvec, ..., CI = TRUE)
}
\arguments{
\item{x}{dataframe, input data}

\item{tvec}{numeric, time on x axis, how much of MWE to plot}

\item{CI}{boolean, plot CIs or not}
}
\description{
Base R line plotting of MWE dataframe with mean, MWE confidence bands, and 1-alpha confidence intervals
}
