% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MWE.R
\name{PQ}
\alias{PQ}
\title{Priority queue}
\usage{
PQ(p)
}
\arguments{
\item{p}{order of column, i.e., output of order (x)}
}
\value{
Lexical closure which contains the following O(1) methods:
 $first(), $second(), $last(), $last2(), and $remove(i) as documented
 in Korpela et al. (2014), appendix A.
}
\description{
Priority queue
}
